Moving SetOfStacks to dynamic array
Changed structure: Improved SetOfStacks memory: now on dynamic array.(controlable size)
Added 2 new fileds: now user can choose number of Stacks in SET and their size using parameter constructor.(added parameter constructor as well).
Added 3 new functions to SetOFstacks:
completelyFull(), completelyEmpty(), topAt(index) to top in certain Stack.
Corrected some sync mistakes
Finished Unit tests for the project!

+++ lab02/dmitru4ok/Stack.h
@@ -32,7 +32,7 @@ public:
 	void clear();               //clear all nodes
 	T pop();                    // return item and remove first node
 	bool empty() const;         // check if is empty
-	size_t size() const;        //count number of list structures
+	int size() const;        //count number of list structures
 	void printStack() const;    //print all elements
 	T top() const;              //watch top element 
 	void emplace(T&);           //== push, but adds the object, not a copy of it
@@ -93,7 +93,7 @@ void Stack<T>::clear()
 }
 
 template <typename T>
-size_t Stack<T>::size() const
+int Stack<T>::size() const
 {
 	if (empty()) return 0;  //if empty size == 0;
 	size_t counter = 1;     //Top already points at Node if not empty, so that = 1;
